Refactor: 이력서작성 저장 DB 연결
