+++ client/package.json
@@ -7,7 +7,7 @@
     "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
-    "axios": "^1.7.2",
+    "axios": "^1.6.1",
     "browser-image-compression": "^2.0.2",
     "dotenv": "^16.4.5",
     "express-session": "^1.18.0",
