+++ server/models/resumes.js
@@ -9,7 +9,6 @@ const resumesSchema = new mongoose.Schema(
     name: { type: String, required: true },
     image: { type: String, default: '' },
     birth: { type: String, required: true },
-    gender: { type: String, required: true },
     phone: { type: String, required: true },
     email: { type: String, required: true },
     address: { type: String, required: true },
